Migrated socket.cpp/hpp into session.cpp/hpp
